MOVE - standard .mle method to Likelihood super class
